add layout component and delete account function
