+++ frontend/src/index.tsx
@@ -3,20 +3,11 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Dashboard from './ContentScreens/Dashboard';
-import Settings from './ContentScreens/Settings';
+
 
 ReactDOM.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<App/>}>
-          <Route index element={<Dashboard/>}/>
-          <Route path="/settings" element={<Settings/>}/>
-        </Route>
-      </Routes>
-    </BrowserRouter>
+    <App/>
   </React.StrictMode>,
   document.getElementById('root')
 );
